Allow clients to configure when a refresh token expires
This allows clients to pass an extra parameter when refreshing a token,
which overrides the configured refresh token timeout in the Synapse
config. This allows a client to opt into a shorter (or longer) lifetime
for their refresh token, which could be used to sign out web sessions
with a specific timeout.

Open questions are mostly if there should be a maximum refresh token
lifetime someone could configure and if this should also be configurable
on login. The latter doesn't seem as necessary, since a client can just
refresh immediately after login (although that is racy).

Once we figure out a nice behaviour for this, we should also write an
MSC. For now this is just an experiment.