chore: add on_upgrade_room() to third_party_event_rules ModuleApi
Called in the process of upgrading a room, provides access to the
`Requester` object for data about the requesting user, and access to the
`RoomVersion` object for data about the requested new version for the
room.

Returns nothing, denials should raise a `SynapseError` with appropriate
codes and messages that will be returned to the client. Multiple modules
may implement this callback, first one that raises an Exception denies
further processing of the upgrade to the room