ci: Basic code coverage setup
Currently only uploads coverage for the unittests run via trial. In
theory we could calculate coverage also from sytest and complement.

The builtin coverage of trial tends to error out when trying to
calculate the coverage of the html files, because it doesn't distinguish
which files got opened. For this reason we are using coverage.py, but
that doesn't work when using multiple trial jobs in parallel (unless
putting in additional work), so this currently uses no concurrency.

I disabled some jobs around the rust parts of synapse for now as those
would take more work to set up.