feat: Allow disabling the faster_join/partial_state room joining mechanism (#35)
This allows an incoming room join request to ignore that the remote
server wished to use a partial_state join. Should also allow the local
server to not request a partial_state join when contacting a remote
server

Disabling is an experimental option, and can be found in the
experimental section of the main Synapse configuration. Defaults to
`True` which allows partial_state joins to occur.

Disable partial_state joins by adding to the shared Synapse
configuration:
```
experimental_features:
  msc3706_enabled: False
```