Bump max backfill denied host count
In the case of an code error response between 400 and 500, a counter is incremented. However,
if the remote server is experiencing a state malfunction it may only think we are not in the
room when we actually are. Perhaps another server in the room disagrees and will disclose room
data for us. Try all the domains until 50 have been tried, then it is more likely of success.