work up enhancing backfill
Two prong approach:
1. Remove the optimization that calculates the backfill request to start
at known backwards extremities. This is technically wrong and needs to
be adjusted slightly still. WIP
2. Adjust the backfill response to include potentially parallel branches
that may exist that would not be normally included in a breadth-first
depth based walk of the DAG. ATM, this only checks the original 'seed'
event that is used in the original request and not 'siblings' of later
events in a branch. Affectively retrieves any forward extremities that
are 'siblings' of the 'seed' event