feature: New docker image tagging for Famedly production builds
This will start producing 3 tagged images for the Famedly docker image
repository. Each image will have the Synapse version encoding and an
appended additional version detail pertaining to the versioned grouping
of modules in a given collection.

This is going to disable the workflow_dispatch system on this workflow
until it can be decided how that should look